CARDS-2742 homepage module: sidebar renamed and import changed, assets renamed (some icons unused)
