+++ modules/homepage/src/main/frontend/assets.config
@@ -1,10 +1,10 @@
 // These lines are aggregated as module.exports entries in webpack.config.js file by the webpack_script.py in aggregated-frontend module
 
-['cards-homepage.themeindex']: './src/themePage/index.jsx'
-['cards-homepage.modelOrganismsIcon']: '@mui/icons-material/Pets'
-['cards-homepage.variantsIcon']: '@mui/icons-material/Subtitles'
-['cards-homepage.adminIcon']: '@mui/icons-material/Settings'
-['cards-homepage.adminDashboard']: './src/adminDashboard/AdminDashboard.jsx'
+['cards-homepage.ThemeIndex']: './src/themePage/index.jsx'
+['cards-homepage.ModelOrganismsIcon']: '@mui/icons-material/Pets'
+['cards-homepage.VariantsIcon']: '@mui/icons-material/Subtitles'
+['cards-homepage.AdminIcon']: '@mui/icons-material/Settings'
+['cards-homepage.AdminDashboard']: './src/adminDashboard/AdminDashboard.jsx'
 ['cards-homepage.QuickSearchResults']: { 'dependOn': ['cards-dataentry.Forms'], 'import': './src/themePage/QuickSearchResults.jsx' }
 ['cards-homepage.QuickSearchConfigurationIcon']: '@mui/icons-material/Pageview'
 ['cards-homepage.QuickSearchConfiguration']: { 'dependOn': ['cards-login.ReLoginDialog'], 'import': './src/themePage/QuickSearchConfiguration' }
