+++ modules/homepage/src/main/resources/SLING-INF/content/Extensions/Views/Administration.json
@@ -2,7 +2,7 @@
   "jcr:primaryType": "cards:Extension",
   "cards:extensionPointId": "cards/coreUI/view",
   "cards:extensionName": "Administration",
-  "cards:extensionRenderURL": "asset:cards-homepage.adminDashboard.js",
+  "cards:extensionRenderURL": "asset:cards-homepage.AdminDashboard.js",
   "cards:targetURL": "/content.html/admin",
   "cards:exactURLMatch": true
 }
