+++ modules/homepage/src/main/resources/SLING-INF/content/libs/cards/Homepage/html.GET.html
@@ -43,6 +43,6 @@
     </script>
     <div id="main-container"></div>
     <sly data-sly-use.assets="/libs/cards/resources/assets">
-      <script src="/libs/cards/resources/${assets['cards-homepage.themeindex.js']}"></script>
+      <script src="/libs/cards/resources/${assets['cards-homepage.ThemeIndex.js']}"></script>
     </sly>
 <sly data-sly-resource="${@selectors='footer'}"/>
